react_search_agent/pyproject.toml
[project]
name = "react-search-agent"
version = "0.1.0"
description = "ReAct-style agent using Tavily search and Groq LLMs"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "langchain~=0.3.0",
    "langchain-core~=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
]